neuro-vitals/backend/requirements.txt
# ── Core FastAPI ──
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
starlette==0.27.0
aiofiles==23.2.1

# ── AI Providers ──
openai==1.3.0
anthropic==0.7.0

# ── Database ──
asyncpg==0.29.0
sqlalchemy==2.0.23
supabase

# ── Utilities ──
python-dateutil==2.8.2
python-dotenv==1.0.0
httpx==0.25.1
requests
pytz

# ── Groq AI (Risk Engine) ──
groq

# ── OCR (Risk Engine - prescription reading) ──
paddleocr
paddlepaddle
pillow
opencv-python-headless
numpy

# ── Data & Visualization ──
pandas
plotly

# ── Google Gemini (Outbreak Analysis) ──
google-generativeai

# ── Google Fit OAuth ──
google-auth
google-auth-oauthlib
google-auth-httplib2